Implemented a left click pickup system with player inventory.
Added a new context callback method for left clicking the mouse that handles what happens to pickup items.
If the camera distance is close enough you can pick up the item.
The item can be destroyed and stored in a scriptable objects list to be used to display the inventory later.