Implemented escape door to escape only when you have the escape key...
that was given after the operation puzzle was complete.
A message will say the door is locked if you click on it without the key and show an end panel where you can go to main menu or restart morgue level and all your inventory will be wiped except items that you brought from the City level.